Feat : fix for non interest bearing debt bug FC-35

+++ src/test/scala/uk/gov/hmrc/test/api/models/sol/SolCalculation.scala
@@ -11,9 +11,9 @@ case class SolCalculation(
   debtId: String,
   mainTrans: String,
   debtTypeDescription: String,
-  interestDueDebtTotal: Int,
-  totalAmountIntDebt: Int,
-  combinedDailyAccrual: Int,
+  interestDueDebtTotal: BigInt,
+  totalAmountIntDebt: BigInt,
+  combinedDailyAccrual: BigInt,
   duties: Seq[SolDuty]
 )
